returning noop when extension is not present (#55)
when extension is not provided for a particular operation,
its not an error for CSM, so we shouldn't return error on it,
instead its noop so CSM will return extension_type as none and
status as none.

For create connection, when no extension is provided, CSM will read
CSM_PARAMETERS env variable, and will return the details by setting
specified list of env variables. If CSM_PARAMTERS is not set then
create connection will also return noop.