dev-mysql: Attempt to auto-detect host part of user name
If the user name configured for the privileged user contains a host name
specification (as defined by having an "@" in the user name),
automatically add it to the (generated) user name.  This makes things
work on Azure, where it refuses to connect to the database unless the
host name is given.